requirements.txt
streamlit
google-cloud-container
google-cloud-compute
google-api-python-client
google-auth
